Refactor: Extract the DPI adjustment logic to a separate file
