+++ src/windows/webview_msedge.h
@@ -131,97 +131,7 @@ private:
   bool m_initialized = false;
 };
 
-// Holds a symbol name and associated type for code clarity.
-template <typename T> class library_symbol {
-public:
-  using type = T;
-
-  constexpr explicit library_symbol(const char *name) : m_name(name) {}
-  constexpr const char *get_name() const { return m_name; }
-
-private:
-  const char *m_name;
-};
-
-// Loads a native shared library and allows one to get addresses for those
-// symbols.
-class native_library {
-public:
-  explicit native_library(const wchar_t *name) : m_handle(LoadLibraryW(name)) {}
-
-  ~native_library() {
-    if (m_handle) {
-      FreeLibrary(m_handle);
-      m_handle = nullptr;
-    }
-  }
-
-  native_library(const native_library &other) = delete;
-  native_library &operator=(const native_library &other) = delete;
-  native_library(native_library &&other) = default;
-  native_library &operator=(native_library &&other) = default;
-
-  // Returns true if the library is currently loaded; otherwise false.
-  operator bool() const { return is_loaded(); }
-
-  // Get the address for the specified symbol or nullptr if not found.
-  template <typename Symbol>
-  typename Symbol::type get(const Symbol &symbol) const {
-    if (is_loaded()) {
-      return reinterpret_cast<typename Symbol::type>(
-          GetProcAddress(m_handle, symbol.get_name()));
-    }
-    return nullptr;
-  }
-
-  // Returns true if the library is currently loaded; otherwise false.
-  bool is_loaded() const { return !!m_handle; }
-
-  void detach() { m_handle = nullptr; }
-
-private:
-  HMODULE m_handle = nullptr;
-};
-
-struct user32_symbols {
-  using DPI_AWARENESS_CONTEXT = HANDLE;
-  using SetProcessDpiAwarenessContext_t = BOOL(WINAPI *)(DPI_AWARENESS_CONTEXT);
-  using SetProcessDPIAware_t = BOOL(WINAPI *)();
-
-  static constexpr auto SetProcessDpiAwarenessContext =
-      library_symbol<SetProcessDpiAwarenessContext_t>(
-          "SetProcessDpiAwarenessContext");
-  static constexpr auto SetProcessDPIAware =
-      library_symbol<SetProcessDPIAware_t>("SetProcessDPIAware");
-};
-
-struct shcore_symbols {
-  typedef enum { PROCESS_PER_MONITOR_DPI_AWARE = 2 } PROCESS_DPI_AWARENESS;
-  using SetProcessDpiAwareness_t = HRESULT(WINAPI *)(PROCESS_DPI_AWARENESS);
-
-  static constexpr auto SetProcessDpiAwareness =
-      library_symbol<SetProcessDpiAwareness_t>("SetProcessDpiAwareness");
-};
-
-inline bool enable_dpi_awareness() {
-  auto user32 = native_library(L"user32.dll");
-  if (auto fn = user32.get(user32_symbols::SetProcessDpiAwarenessContext)) {
-    if (fn(DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE)) {
-      return true;
-    }
-    return GetLastError() == ERROR_ACCESS_DENIED;
-  }
-  if (auto shcore = native_library(L"shcore.dll")) {
-    if (auto fn = shcore.get(shcore_symbols::SetProcessDpiAwareness)) {
-      auto result = fn(shcore_symbols::PROCESS_PER_MONITOR_DPI_AWARE);
-      return result == S_OK || result == E_ACCESSDENIED;
-    }
-  }
-  if (auto fn = user32.get(user32_symbols::SetProcessDPIAware)) {
-    return !!fn();
-  }
-  return true;
-}
+#include "mswebview_engine.hpp"
 
 // Enable built-in WebView2Loader implementation by default.
 #ifndef WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL
